Exercise 12.2.1 add word_to_tuple function and anagram function
Add a function to convert a given word to a tuple; abstracted to it's own function for good program design.

Add the logic for the anagram function; turns out the whole thing was a lot simpler than I thought it would be, so I can do most of the program in one commit.